Moves images to their own directory
